+++ packages/tracing/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## 0.4.0 - 2020-08-20
 
-- The exported `Span` interface is now extended from the one provided by `@opentelemetry/api`.
+- The exported `Span` interface is now different from the one provided by `@opentelemetry/api` and only exposees `setAttribute`, `setDisclosedAttribute` and `end`.
 
-- When using `setAttribute` or `setAttributes` on a `Span`, it will obfuscate values by default. Use `setDisclosedAttribute` or `setDisclosedAttributes` to explicitely do so for a value.
+- When using `setAttribute` on a `Span`, it will obfuscate values by default. Use `setDisclosedAttribute` to explicitely do so for a value.
 
 
 ## 0.3.3 - 2020-08-19